requirements.txt
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
numpy==2.2.6
pandas==2.3.0
pyaes==1.6.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
soupsieve==2.7
Telethon==1.40.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
